refactor(cs-bridge): merge cli.ts into index.ts
Move setupRegistration, startBridge, runCli into index.ts.
Delete cli.ts.

Phase 4.4 of kernel philosophy refactoring (5 files achieved).
